Let the padding be generated by `ID3.serialize()`.
Instead of generating the padding in the `ID3.to_file()` method, it is the
responsibility of the `ID3.serialize()` method to append a proper padding. To do
this, the already existing optional argument `min_length` is used to ensure that
`ID3.serialize()` adds a padding making the tag at least as long as the tag
currently present in the file.